requirements.txt
# Core web framework
fastapi==0.115.0
uvicorn[standard]==0.30.1
gunicorn==22.0.0

# NLP and ML libraries
transformers==4.46.1
torch==2.5.1

# Web scraping and article extraction
beautifulsoup4==4.12.3
requests==2.32.3
newspaper3k==0.2.8

# Frontend templating
jinja2==3.1.4
python-multipart==0.0.9

# Utilities
httpx==0.27.0
lxml==5.3.0
